Remove cpp, python, and cmake files
