+++ examples/qualcomm/CMakeLists.txt
@@ -80,12 +80,6 @@ add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/oss_scripts/t5)
 # build qnn_whisper_runner for whisper
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/oss_scripts/whisper)
 
-# build qaihub_llama2_7b_runner and qaihub_llama3_8b_runner
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/qaihub_scripts/llama)
-
-# build qaihub_stable_diffusion_runner
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/qaihub_scripts/stable_diffusion)
-
 # direct-mode qnn_executor_direct_runner is build with ndk toolchain, paired
 # with libraries built with hexagon toolchain, communicate through self-defined
 # fastrpc protocol.
