--- examples/qualcomm/qaihub_scripts/stable_diffusion/install_requirements.sh
@@ -1,3 +1,0 @@
-# For Stable Diffusion V2.1
-pip install diffusers==0.29.0
-pip install piq==0.8.0
