+++ examples/qualcomm/util_scripts/cli.py
@@ -51,7 +51,7 @@ from executorch.backends.qualcomm.utils.utils import (
     to_edge_transform_and_lower_to_qnn,
 )
 from executorch.devtools import Inspector
-from executorch.examples.qualcomm.qaihub_scripts.utils.utils import preprocess_binary
+from executorch.backends.qualcomm.export_utils import preprocess_binary
 from executorch.exir import ExecutorchBackendConfig
 from executorch.exir.passes.memory_planning_pass import MemoryPlanningPass
 from torchao.quantization import pt2e
@@ -73,6 +73,13 @@ def get_logger():
     logger.setLevel(logging.INFO)
     logger.propagate = False
     return logging.LoggerAdapter(logger, extra={"prefix": "QNN_BACKEND"})
+
+
+def preprocess_binary(ctx_bin, compiler_specs):
+    qnn_mgr = PyQnnManagerAdaptor.QnnManager(
+        generate_qnn_executorch_option(compiler_specs),
+    )
+    return bytes(qnn_mgr.MakeBinaryInfo(ctx_bin))
 
 
 def get_io_info(pte_path, compiler_specs):
